requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
click==8.1.6
dnspython==2.4.0
fastapi==0.100.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
pydantic==2.0.3
pydantic_core==2.3.0
PyJWT==2.8.0
pymongo==4.4.1
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
Werkzeug==2.3.6
